refactor(pod): extract shared length-prefix module
PodString and PodVec both had identical copies of max_n_for_pfx,
decode/encode prefix helpers, and the _CAP_CHECK assertion block. Move
them into a new prefix module and re-export via a cap_check! macro.

Net deletion of ~40 lines of duplicated logic with no behavioral change.
