+++ pod/src/string.rs
@@ -42,20 +42,10 @@
 //! loaded on guard creation, flushed back (with one realloc CPI if size
 //! changes) on drop.
 
-use core::mem::MaybeUninit;
-
-/// Returns the maximum `N` value representable by a `PFX`-byte length prefix.
-///
-/// Returns `0` for invalid `PFX` values, which causes `_CAP_CHECK` to fire.
-pub(crate) const fn max_n_for_pfx(pfx: usize) -> usize {
-    match pfx {
-        1 => u8::MAX as usize,
-        2 => u16::MAX as usize,
-        4 => u32::MAX as usize,
-        8 => usize::MAX,
-        _ => 0,
-    }
-}
+use {
+    crate::prefix::{cap_check, decode_prefix_len, encode_prefix_len},
+    core::mem::MaybeUninit,
+};
 
 /// Fixed-capacity inline string stored in account data.
 ///
@@ -79,17 +69,7 @@ pub struct PodString<const N: usize, const PFX: usize = 1> {
 
 // Compile-time: PFX must be in {1,2,4,8} and N must fit in the prefix.
 impl<const N: usize, const PFX: usize> PodString<N, PFX> {
-    const _CAP_CHECK: () = {
-        assert!(
-            PFX == 1 || PFX == 2 || PFX == 4 || PFX == 8,
-            "PodString<N, PFX>: PFX must be 1, 2, 4, or 8"
-        );
-        assert!(
-            N <= max_n_for_pfx(PFX),
-            "PodString<N, PFX>: N exceeds the maximum value representable by the PFX-byte length \
-             prefix"
-        );
-    };
+    const _CAP_CHECK: () = cap_check!("PodString<N, PFX>", N, PFX);
 
     /// Compile-time validity check. Reference this in a `const` context to
     /// verify that `N` and `PFX` are in range at the call site.
@@ -122,16 +102,11 @@ const _: () = assert!(core::mem::align_of::<PodString<0, 8>>() == 1);
 
 impl<const N: usize, const PFX: usize> PodString<N, PFX> {
     /// Decode the on-disk length prefix into a `usize`.
-    ///
-    /// LLVM constant-folds this per monomorphization (e.g., for PFX=1 it
-    /// compiles to a single byte load).
     #[inline(always)]
     fn decode_len(&self) -> usize {
         #[allow(clippy::let_unit_value)]
         let _ = Self::_CAP_CHECK;
-        let mut buf = [0u8; 8];
-        buf[..PFX].copy_from_slice(&self.len);
-        u64::from_le_bytes(buf) as usize
+        decode_prefix_len(&self.len)
     }
 
     /// Encode `n` as a `PFX`-byte little-endian prefix into `self.len`.
@@ -139,8 +114,7 @@ impl<const N: usize, const PFX: usize> PodString<N, PFX> {
     fn encode_len(&mut self, n: usize) {
         #[allow(clippy::let_unit_value)]
         let _ = Self::_CAP_CHECK;
-        let bytes = (n as u64).to_le_bytes();
-        self.len.copy_from_slice(&bytes[..PFX]);
+        encode_prefix_len(&mut self.len, n);
     }
 
     /// Number of active bytes in the string.
@@ -277,9 +251,11 @@ impl<const N: usize, const PFX: usize> PodString<N, PFX> {
             bytes.len() >= PFX,
             "load_from_bytes: slice must have at least PFX bytes"
         );
-        let mut buf = [0u8; 8];
-        buf[..PFX].copy_from_slice(&bytes[..PFX]);
-        let slen = (u64::from_le_bytes(buf) as usize).min(N);
+        // SAFETY: debug_assert above ensures bytes.len() >= PFX. We need a
+        // [u8; PFX] reference for decode_prefix_len; try_into is infallible
+        // given the length check.
+        let pfx_bytes: &[u8; PFX] = bytes[..PFX].try_into().unwrap();
+        let slen = decode_prefix_len(pfx_bytes).min(N);
         debug_assert!(
             bytes.len() >= PFX + slen,
             "load_from_bytes: slice too short for encoded length"
